+++ example/lib/external_account_alias.dart
@@ -4,5 +4,6 @@ import 'account.dart';
 
 class Alias {}
 
+// This can be useful for when the types are defined in a package outside of your code.
 @JObj()
 class ExternalAccountAlias = Account with Alias;